refactor(qwen3): move the report harness out of the library API
`pegainfer-qwen3` exposed `pub mod kernel_bench` (1112 lines) as public
API, so every normal library and server build compiled report-only
benchmark code and shipped its types in the crate's public surface.

Move the harness to `src/bin/report_support/`, split across five modules,
and `#[path]`-include it from the two `kernel-report` binaries. It is now
compiled per binary and never as part of the library.

- `qwen3_kernel_report` takes the whole tree; `qwen3_model_report` needs
  only `common::L2CacheClear` and includes that module alone, with the
  `#[allow(dead_code)]` on its own `mod` declaration so the other four
  modules stay dead-code-checked by the kernel report's build.
- The model report now imports production primitives from their owning
  crates: `build_split_kv_csr` from `pegainfer_core::ops` and
  `split_qkv_into` from `pegainfer_kernels::ops`, instead of reaching
  through a harness re-export. This also fixes a pre-existing build
  break: `qwen3_model_report` did not compile under `--features
  kernel-report` at HEAD, because `split_qkv_into` is not re-exported by
  `pegainfer-core`.
- `SplitKvConfig::{new, label}` widen to `pub`. `parse` was already
  public and calls `new`, so a private constructor only pushed external
  callers to the unchecked struct-literal path; `label` is `parse`'s
  inverse and restating its format in the harness would silently break
  the round-trip contract.
- `SPLIT_KV_CHUNK_TOKENS` / `SPLIT_KV_TUNED_MAX_CHUNKS` are re-exported
  through `runtime` so the harness keeps deriving the Tuned split width
  from the runtime constants, as it did before the move, rather than
  restating 64x64 where a retune would silently desync `split_tmp_*`.

Report schemas, CLI flags, benchmark cases, numeric policy, and kernel
selection are unchanged; both binaries' `--help` output is identical.

Verified on RTX 5090 (sm_120):
  cargo check  -p pegainfer-qwen3 --lib                                       0 warnings
  cargo check  -p pegainfer-qwen3 --all-targets --features kernel-report      pass
  cargo clippy -p pegainfer-qwen3 --all-targets --features kernel-report -D warnings  pass
  cargo test   -p pegainfer-qwen3 --lib                                       88 passed
  cargo build  --features kernel-report --bin qwen3_kernel_report --bin qwen3_model_report  both link
  cargo fmt --check                                                           pass

Closes #944
